implment counter + monthly stats
